add(src): Ентити эффекты галлюцинаций
