--- Content.Shared/ADT/Hallucinations/HallucinationsEffect.cs
@@ -1,37 +1,0 @@
-using Robust.Shared.Prototypes;
-
-namespace Content.Shared.EntityEffects.Effects
-{
-    /// <summary>
-    /// Default metabolism for stimulants and tranqs. Attempts to find a MovementSpeedModifier on the target,
-    /// adding one if not there and to change the movespeed
-    /// </summary>
-    public sealed partial class HallucinationsReagentEffect : EntityEffectBase<HallucinationsReagentEffect>
-    {
-        [DataField(required: true)]
-        public List<string> Hallucinations;
-
-        [DataField]
-        public float Time = 2.0f;
-
-        [DataField]
-        public bool Refresh = true;
-
-        [DataField]
-        public HallucinationsMetabolismType Type = HallucinationsMetabolismType.Add;
-
-        public override string? EntityEffectGuidebookText(IPrototypeManager prototype, IEntitySystemManager entSys)
-        {
-            return Loc.GetString("reagent-effect-guidebook-hallucinations",
-                ("chance", Probability),
-                ("time", Time));
-        }
-    }
-
-    public enum HallucinationsMetabolismType
-    {
-        Add,
-        Remove,
-        Set
-    }
-}
